Revert "Product DLS Validation"
