+++ CSV/errors.go
@@ -1,0 +1,7 @@
+package csv
+
+import "errors"
+
+var (
+	errNoData = errors.New("No data present")
+)
