--- pride_priderelease_20221122164529.csv
@@ -1,5 +1,0 @@
-CatalogueOrder,BrandscopeCarryOver,Integration_ID,SKU,Barcode,ProductName,ProductColourCode,ProductDisplayColour,GenericColour,SizeBreak,AttributeValue,AttributeType,AttributeSequence,DisplayWholesale,DisplayWholesaleRange,WholesalePriceOriginal,WholesalePrice,DisplayRetail,RetailPriceOriginal,RetailPrice,PackUnits,ProductMultiple,AvailableMonths,Divisions,Collections,Categories,DiscountCategory,ProductTags,AgeGroup,Gender,BrandscopeHierarchy,State,ProductRequirement1,ProductSpecification1,ProductSpecification2,ProductSpecification3,ProductSpecification4,ProductSpecification5,ProductSpecification6,ProductSpecification7,ProductSpecification8,ProductSpecification9,ProductSpecification10,ProductSpecification11,ProductSpecification12,ProductSpecification13,ProductSpecification14,ProductSpecification15,ProductChanges1,ProductChanges2,ProductChanges3,ProductChanges4,ProductChanges5,AdditionalDetail1,AdditionalDetail2,AdditionalDetail3,AdditionalDetail4,AdditionalDetail5,SalesTip,MarketingSupport,CompanyName,BrandName,ReleaseName,SegmentNames,AtsInIndent,AtsInInseason,PreOrderLeadTimeDays,PreOrderMessage,status,ok
-12,N,as12,5345,7788hh,Gucci,as12,Black,Black,XL,,as,,12,,76,2,76,10,10,1,0,SEP22,Unassigned,,Drybag,,,Adult,Male,a,Active,Awesome Collection,asqw,,,,,,,,,,,,,,,,,,,,Details,,,,,,,clop12,pride,priderelease,,N,Y,,
-12,N,as12,5345,7788hh,Gucci,as12,Black,Black,XL,,as,,12,,76,2,76,10,10,1,0,SEP22,Unassigned,,Drybag,,,Adult,Male,a,Active,Awesome Collection,asqw,,,,,,,,,,,,,,,,,,,,Details,,,,,,,clop12,pride,priderelease,,N,Y,,
-12,N,as12,5345,7788hh,Gucci,as12,Black,Black,XL,,as,,12,,76,2,76,10,10,1,0,SEP22,Unassigned,,Drybag,,,Adult,Male,a,Active,Awesome Collection,asqw,,,,,,,,,,,,,,,,,,,,Details,,,,,,,clop12,pride,priderelease,,N,Y,,
-12,N,as12,5345,7788hh,Gucci,as12,Black,Black,XL,,as,,12,,76,2,76,10,10,1,0,SEP22,Unassigned,,Drybag,,,Adult,Male,a,Active,Awesome Collection,asqw,,,,,,,,,,,,,,,,,,,,Details,,,,,,,clop12,pride,priderelease,,N,Y,,
